+++ paramiko/client.py
@@ -26,8 +26,6 @@ import inspect
 import os
 import socket
 import warnings
-from errno import ECONNREFUSED, EHOSTUNREACH
-
 from paramiko.agent import Agent
 from paramiko.common import DEBUG
 from paramiko.config import SSH_PORT
@@ -39,7 +37,6 @@ from paramiko.rsakey import RSAKey
 from paramiko.ssh_exception import (
     SSHException,
     BadHostKeyException,
-    NoValidConnectionsError,
 )
 from paramiko.transport import Transport
 from paramiko.util import ClosingContextManager
@@ -190,30 +187,6 @@ class SSHClient(ClosingContextManager):
             policy = policy()
         self._policy = policy
 
-    def _families_and_addresses(self, hostname, port):
-        """
-        Yield pairs of address families and addresses to try for connecting.
-
-        :param str hostname: the server to connect to
-        :param int port: the server port to connect to
-        :returns: Yields an iterable of ``(family, address)`` tuples
-        """
-        guess = True
-        addrinfos = socket.getaddrinfo(
-            hostname, port, socket.AF_UNSPEC, socket.SOCK_STREAM
-        )
-        for (family, socktype, proto, canonname, sockaddr) in addrinfos:
-            if socktype == socket.SOCK_STREAM:
-                yield family, sockaddr
-                guess = False
-
-        # some OS like AIX don't indicate SOCK_STREAM support, so just
-        # guess. :(  We only do this if we did not get a single result marked
-        # as socktype == SOCK_STREAM.
-        if guess:
-            for family, _, _, _, sockaddr in addrinfos:
-                yield family, sockaddr
-
     def connect(
         self,
         hostname,
@@ -343,41 +316,7 @@ class SSHClient(ClosingContextManager):
             Added the ``transport_factory`` argument.
         """
         if not sock:
-            errors = {}
-            # Try multiple possible address families (e.g. IPv4 vs IPv6)
-            to_try = list(self._families_and_addresses(hostname, port))
-            for af, addr in to_try:
-                try:
-                    sock = socket.socket(af, socket.SOCK_STREAM)
-                    if timeout is not None:
-                        try:
-                            sock.settimeout(timeout)
-                        except:
-                            pass
-                    sock.connect(addr)
-                    # Break out of the loop on success
-                    break
-                except socket.error as e:
-                    # As mentioned in socket docs it is better
-                    # to close sockets explicitly
-                    if sock:
-                        sock.close()
-                    # Raise anything that isn't a straight up connection error
-                    # (such as a resolution error)
-                    if e.errno not in (ECONNREFUSED, EHOSTUNREACH):
-                        raise
-                    # Capture anything else so we know how the run looks once
-                    # iteration is complete. Retain info about which attempt
-                    # this was.
-                    errors[addr] = e
-
-            # Make sure we explode usefully if no address family attempts
-            # succeeded. We've no way of knowing which error is the "right"
-            # one, so we construct a hybrid exception containing all the real
-            # ones, of a subclass that client code should still be watching for
-            # (socket.error)
-            if len(errors) == len(to_try):
-                raise NoValidConnectionsError(errors)
+            sock = socket.create_connection((hostname, port), timeout=timeout)
 
         if transport_factory is None:
             transport_factory = Transport